requirements.txt
fastapi==0.104.1
gunicorn==21.2.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.1
python-multipart==0.0.6
requests==2.31.0
scipy==1.11.4
numpy==1.26.2
librosa==0.10.0
soundfile==0.12.1
